requirements.txt
fastapi
uvicorn[standard]
python-multipart
jinja2
websockets
moviepy
pillow
git+https://github.com/jebin2/STT.git#egg=stt-runner[fasterwhisper]
git+https://github.com/jebin2/custom_logger.git